==> code/03_word2vec_immigyears.ipynb <==
import glob
import re
import gensim
import copy
import numpy
import pandas as pd
import pickle
import random
from statistics import stdev
from gensim.models import Word2Vec
from sklearn.utils import resample

random.seed(91326)

model = Word2Vec.load("intermediate/immigration.model")
keys = model.wv.key_to_index
rwords = [key for key, index in keys.items()]

similarities = {}
for word in rwords:
    similarity_score = model.wv.similarity("immigration", word)
    similarities[word] = similarity_score

wordsmean = sum(similarities.values())/len(similarities)
std = stdev(similarities.values())
maxi = max(similarities.values()) 
mini = min(similarities.values())
print(wordsmean, std, maxi, mini)

def chrono_train_immig(n_iterations, current_corpus, previous_model, output_model):
    ''' Models the current corpus by initializing with the vectors of  '''
    ''' the previous model, outputs similarity scores of new model,    '''
    ''' and saves that new model for the next round of modeling        '''
    racialization = []
    latinx = []
    asian = []
    black = []
    wall = []
    placebo = []

    tempmodel = Word2Vec.load(previous_model)

    for k in range(n_iterations):
        model = tempmodel
        sentence_samples = resample(current_corpus)
        model.train(sentence_samples, total_examples = len(sentence_samples), epochs = 1)
        racialization.append(model.wv.similarity('immigration','racialization'))
        latinx.append(model.wv.similarity('immigration', 'latinxdad'))
        asian.append(model.wv.similarity('immigration', 'azn'))
        black.append(model.wv.similarity('immigration', 'blck'))
        wall.append(model.wv.similarity('immigration', 'wall'))

        nums = random.sample(range(1,len(model.wv.key_to_index)), 100)
        keys = model.wv.key_to_index
        rwords = [key for key, index in keys.items() if index in nums]
        similarities = {}
        for word in rwords:
            similarity_score = model.wv.similarity("healthcare", word)
            similarities[word] = similarity_score
    
        rwordsmean = sum(similarities.values())/len(similarities)
        placebo.append(rwordsmean)
        
        run = k+1
        print("Finished with run %d out of %d" % (run, n_iterations))
    model.save(output_model)
    stats = racialization, latinx, asian, black, wall, placebo
    return list(stats) 


eras = ['2004', '2005', '2006', '2007', '2008', '2009', '2010', '2011', '2012', '2013', '2014', '2015', '2016', '2017', '2018', '2019', '2020', '2021', '2022']

results_all = []

with open('intermediate/immigration_corpus.pickle', 'rb') as f:
    list_of_lists = pickle.load(f)

results_2004 = chrono_train_immig(100, list_of_lists[0], "intermediate/immigration.model", "intermediate/word2vec_models/immigration_model1_of_2004.model")
results_all.append(results_2004)

results_2005 = chrono_train_immig(100, list_of_lists[1], "intermediate/word2vec_models/immigration_model1_of_2004.model", "intermediate/word2vec_models/immigration_model2_of_2005.model")
results_all.append(results_2005)

results_2006 = chrono_train_immig(100, list_of_lists[2], "intermediate/word2vec_models/immigration_model2_of_2005.model", "intermediate/word2vec_models/immigration_model3_of_2006.model")
results_all.append(results_2006)

results_2007 = chrono_train_immig(100, list_of_lists[3], "intermediate/word2vec_models/immigration_model3_of_2006.model", "intermediate/word2vec_models/immigration_model4_of_2007.model")
results_all.append(results_2007)

results_2008 = chrono_train_immig(100, list_of_lists[4], "intermediate/word2vec_models/immigration_model4_of_2007.model", "intermediate/word2vec_models/immigration_model5_of_2008.model")
results_all.append(results_2008)

results_2009 = chrono_train_immig(100, list_of_lists[5], "intermediate/word2vec_models/immigration_model5_of_2008.model", "intermediate/word2vec_models/immigration_model6_of_2009.model")
results_all.append(results_2009)

results_2010 = chrono_train_immig(100, list_of_lists[6], "intermediate/word2vec_models/immigration_model6_of_2009.model", "intermediate/word2vec_models/immigration_model7_of_2010.model")
results_all.append(results_2010)

results_2011 = chrono_train_immig(100, list_of_lists[7], "intermediate/word2vec_models/immigration_model7_of_2010.model", "intermediate/word2vec_models/immigration_model8_of_2011.model")
results_all.append(results_2011)

results_2012 = chrono_train_immig(100, list_of_lists[8], "intermediate/word2vec_models/immigration_model8_of_2011.model", "intermediate/word2vec_models/immigration_model9_of_2012.model")
results_all.append(results_2012)

results_2013 = chrono_train_immig(100, list_of_lists[9], "intermediate/word2vec_models/immigration_model9_of_2012.model", "intermediate/word2vec_models/immigration_model10_of_2013.model")
results_all.append(results_2013)

results_2014 = chrono_train_immig(100, list_of_lists[10], "intermediate/word2vec_models/immigration_model10_of_2013.model", "intermediate/word2vec_models/immigration_model11_of_2014.model")
results_all.append(results_2014)

results_2015 = chrono_train_immig(100, list_of_lists[11], "intermediate/word2vec_models/immigration_model11_of_2014.model", "intermediate/word2vec_models/immigration_model12_of_2015.model")
results_all.append(results_2015)

results_2016 = chrono_train_immig(100, list_of_lists[12], "intermediate/word2vec_models/immigration_model12_of_2015.model", "intermediate/word2vec_models/immigration_model13_of_2016.model")
results_all.append(results_2016)

results_2017 = chrono_train_immig(100, list_of_lists[13], "intermediate/word2vec_models/immigration_model13_of_2016.model", "intermediate/word2vec_models/immigration_model14_of_2017.model")
results_all.append(results_2017)

results_2018 = chrono_train_immig(100, list_of_lists[14], "intermediate/word2vec_models/immigration_model14_of_2017.model", "intermediate/word2vec_models/immigration_model15_of_2018.model")
results_all.append(results_2018)

results_2019 = chrono_train_immig(100, list_of_lists[15], "intermediate/word2vec_models/immigration_model15_of_2018.model", "intermediate/word2vec_models/immigration_model16_of_2019.model")
results_all.append(results_2019)

results_2020 = chrono_train_immig(100, list_of_lists[16], "intermediate/word2vec_models/immigration_model16_of_2019.model", "intermediate/word2vec_models/immigration_model17_of_2020.model")
results_all.append(results_2020)

results_2021 = chrono_train_immig(100, list_of_lists[17], "intermediate/word2vec_models/immigration_model17_of_2020.model", "intermediate/word2vec_models/immigration_model18_of_2021.model")
results_all.append(results_2021)

results_2022 = chrono_train_immig(100, list_of_lists[18], "intermediate/word2vec_models/immigration_model18_of_2021.model", "intermediate/word2vec_models/immigration_model19_of_2022.model")
results_all.append(results_2022)

racialization_similarity = []
latinx_similarity = []
asian_similarity = []
black_similarity = []
wall_similarity = []
placebo_similarity = []

for i in range(0, len(results_all)):
    racialization = results_all[i][0]
    racialization_similarity.append(racialization)
    latinx = results_all[i][1]
    latinx_similarity.append(latinx)
    asian = results_all[i][2]
    asian_similarity.append(asian)
    black = results_all[i][3]
    black_similarity.append(black)
    wall = results_all[i][4]
    wall_similarity.append(wall)
    placebo = results_all[i][5]
    placebo_similarity.append(placebo)
    
stat_types = [racialization_similarity, latinx_similarity, asian_similarity, black_similarity, wall_similarity, placebo_similarity]

## Saving model output

with open('aggregate_models/immigration_chrono_model_output.pickle', 'wb') as f:
    pickle.dump(stat_types, f)
