==> code/04_estimates_hc.ipynb <==
import numpy
import pickle
import csv
import os
import math
import random
import pandas as pd
from gensim.models import Word2Vec

# Healthcare 2004-2022 Script

eras = ['2004', '2005', '2006', '2007', '2008', '2009', '2010', '2011', '2012', '2013', '2014', '2015', '2016', '2017', '2018', '2019', '2020', '2021', '2022']

with open("aggregate_models/healthcare_chrono_model_output.pickle", 'rb') as f:
    stat_types = pickle.load(f)               
    
## Finding means

means_wCI = pd.DataFrame(columns=['mean', 'ci low', 'ci high'])

for t in stat_types:
    stats = []
    for e in range(0, len(t)):
        test = [score for score in t[e] if score != 'NA']
        test = numpy.asarray(test)
        test_mean = test[0:99].mean()                   # Specify the number of samples to use
        test_error = 1.96 * (test.std() / math.sqrt(100))  # Specify the number of samples you used
        test_lowci = test_mean - test_error
        test_highci = test_mean + test_error
        test_stats = [test_mean, test_lowci, test_highci]
        stats.append(test_stats)
    means_wCI = pd.concat([means_wCI, pd.DataFrame(stats, columns=['mean', 'ci low', 'ci high'])], ignore_index=True)

partition = len(means_wCI) // len(stat_types)    
policy = ['racialization', 'obamacare', 'obama', 'gender', 'poverty', 'black', 'placebo']
means_wCI['type'] = numpy.repeat(policy, partition)

## Exporting as a .csv file for visualization in R

means_wCI['year'] = eras*len(stat_types)

means_wCI.to_csv('healthcare_chrono_means_wCI.csv')
