==> code/05_modeleval.ipynb <==
import numpy
import pickle
import csv
import os
import math
import random
import pandas as pd
from gensim.models import Word2Vec
from gensim.models import KeyedVectors
from gensim.test.utils import datapath

# Loading in policy baseline models

with open("intermediate/healthcare.model", 'rb') as f:
    healthcare = pickle.load(f)   

with open("intermediate/immigration.model", 'rb') as f:
    immigration = pickle.load(f)
    
with open("intermediate/crime.model", 'rb') as f:
    crime = pickle.load(f)

hc_ws353 = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-WS-353-ALL.txt')[0]
hc_wssim = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-WS-353-SIM.txt')[0]
hc_wsrel = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-WS-353-REL.txt')[0]
hc_mc = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-MC-30.txt')[0]
hc_rg = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-RG-65.txt')[0]
hc_rw = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-RW-STANFORD.txt')[0]
hc_men = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-MEN-TR-3k.txt')[0]
hc_mturk287 = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-MTurk-287.txt')[0]
hc_mturk771 = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-MTurk-771.txt')[0]
hc_yp = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-YP-130.txt')[0]
hc_simlex = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-SIMLEX-999.txt')[0]
hc_verb = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-VERB-143.txt')[0]
hc_simverb = healthcare.wv.evaluate_word_pairs('intermediate/analogy/EN-SimVerb-3500.txt')[0]

immig_ws353 = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-WS-353-ALL.txt')[0]
immig_wssim = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-WS-353-SIM.txt')[0]
immig_wsrel = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-WS-353-REL.txt')[0]
immig_mc = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-MC-30.txt')[0]
immig_rg = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-RG-65.txt')[0]
immig_rw = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-RW-STANFORD.txt')[0]
immig_men = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-MEN-TR-3k.txt')[0]
immig_mturk287 = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-MTurk-287.txt')[0]
immig_mturk771 = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-MTurk-771.txt')[0]
immig_yp = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-YP-130.txt')[0]
immig_simlex = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-SIMLEX-999.txt')[0]
immig_verb = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-VERB-143.txt')[0]
immig_simverb = immigration.wv.evaluate_word_pairs('intermediate/analogy/EN-SimVerb-3500.txt')[0]

crime_ws353 = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-WS-353-ALL.txt')[0]
crime_wssim = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-WS-353-SIM.txt')[0]
crime_wsrel = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-WS-353-REL.txt')[0]
crime_mc = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-MC-30.txt')[0]
crime_rg = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-RG-65.txt')[0]
crime_rw = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-RW-STANFORD.txt')[0]
crime_men = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-MEN-TR-3k.txt')[0]
crime_mturk287 = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-MTurk-287.txt')[0]
crime_mturk771 = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-MTurk-771.txt')[0]
crime_yp = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-YP-130.txt')[0]
crime_simlex = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-SIMLEX-999.txt')[0]
crime_verb = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-VERB-143.txt')[0]
crime_simverb = crime.wv.evaluate_word_pairs('intermediate/analogy/EN-SimVerb-3500.txt')[0]

# Evaluating analogies for each policy baseline model 

hc_analogy_scores = healthcare.wv.evaluate_word_analogies(datapath('questions-words.txt'))[0]

immig_analogy_scores = immigration.wv.evaluate_word_analogies(datapath('questions-words.txt'))[0]

crime_analogy_scores = crime.wv.evaluate_word_analogies(datapath('questions-words.txt'))[0]

# Exporting 

# Sample data for six objects
objects = [
    {"name": "hc_analogy", "value": hc_analogy_scores},
    {"name": "immig_analogy", "value": immig_analogy_scores},
    {"name": "crime_analogy", "value": crime_analogy_scores},
    {"name": "hc_ws353", "value": hc_ws353},
    {"name": "hc_wssim", "value": hc_wssim},
    {"name": "hc_wsrel", "value": hc_wsrel},
    {"name": "hc_mc", "value": hc_mc},
    {"name": "hc_rg", "value": hc_rg},
    {"name": "hc_rw", "value": hc_rw},
    {"name": "hc_men", "value": hc_men},
    {"name": "hc_mturk287", "value": hc_mturk287},
    {"name": "hc_mturk771", "value": hc_mturk771},
    {"name": "hc_yp", "value": hc_yp},
    {"name": "hc_simlex", "value": hc_simlex},
    {"name": "hc_verb", "value": hc_verb},
    {"name": "hc_simverb", "value": hc_simverb},
    {"name": "immig_ws353", "value": immig_ws353},
    {"name": "immig_wssim", "value": immig_wssim},
    {"name": "immig_wsrel", "value": immig_wsrel},
    {"name": "immig_mc", "value": immig_mc},
    {"name": "immig_rg", "value": immig_rg},
    {"name": "immig_rw", "value": immig_rw},
    {"name": "immig_men", "value": immig_men},
    {"name": "immig_mturk287", "value": immig_mturk287},
    {"name": "immig_mturk771", "value": immig_mturk771},
    {"name": "immig_yp", "value": immig_yp},
    {"name": "immig_simlex", "value": immig_simlex},
    {"name": "immig_verb", "value": immig_verb},
    {"name": "immig_simverb", "value": immig_simverb},
    {"name": "crime_ws353", "value": crime_ws353},
    {"name": "crime_wssim", "value": crime_wssim},
    {"name": "crime_wsrel", "value": crime_wsrel},
    {"name": "crime_mc", "value": crime_mc},
    {"name": "crime_rg", "value": crime_rg},
    {"name": "crime_rw", "value": crime_rw},
    {"name": "crime_men", "value": crime_men},
    {"name": "crime_mturk287", "value": crime_mturk287},
    {"name": "crime_mturk771", "value": crime_mturk771},
    {"name": "crime_yp", "value": crime_yp},
    {"name": "crime_simlex", "value": crime_simlex},
    {"name": "crime_verb", "value": crime_verb},
    {"name": "crime_simverb", "value": crime_simverb},
]

# Define the CSV file path
csv_file_path = "eval_models.csv"

# Define the headers for CSV
headers = ["name", "value"]

# Write data to CSV file
with open(csv_file_path, "w", newline="") as csvfile:
    writer = csv.DictWriter(csvfile, fieldnames=headers)
    writer.writeheader()
    for obj in objects:
        writer.writerow(obj)
