==> code/07_nn_race.ipynb <==
import pandas as pd
from gensim.models import Word2Vec

def get_nearest_neighbors(model, word, topn=10):
    try:
        neighbors = model.wv.most_similar(positive=[word], topn=topn)
        return [neighbor[0] for neighbor in neighbors]
    except KeyError:
        print(f"Word '{word}' not in vocabulary.")
        return []

def extract_neighbors_to_dataframe(model, word):
    neighbors = get_nearest_neighbors(model, word)
    df = pd.DataFrame({'word': [word], 'neighbors': [', '.join(neighbors)]})
    return df

# Load your Word2Vec model
model = Word2Vec.load("intermediate/immigration.model")

# Extract neighbors and store in DataFrame
neighbors_df_word1 = extract_neighbors_to_dataframe(model, "racialization")
neighbors_df_word2 = extract_neighbors_to_dataframe(model, "blck")
neighbors_df_word3 = extract_neighbors_to_dataframe(model, "azn")
neighbors_df_word4 = extract_neighbors_to_dataframe(model, "latinxdad")

# Concatenate DataFrames
combined_df = pd.concat([neighbors_df_word1, neighbors_df_word2, neighbors_df_word3, neighbors_df_word4], ignore_index=True)

# Save the combined DataFrame to a CSV file
combined_df.to_csv("immig_neighbors_race.csv", index=False)

print("DataFrame saved to 'word_neighbors.csv'")


# Load your Word2Vec model
model = Word2Vec.load("intermediate/crime.model")

# Extract neighbors and store in DataFrame
neighbors_df_word1 = extract_neighbors_to_dataframe(model, "racialization")
neighbors_df_word2 = extract_neighbors_to_dataframe(model, "blck")
neighbors_df_word3 = extract_neighbors_to_dataframe(model, "azn")
neighbors_df_word4 = extract_neighbors_to_dataframe(model, "latinxdad")

# Concatenate DataFrames
combined_df = pd.concat([neighbors_df_word1, neighbors_df_word2, neighbors_df_word3, neighbors_df_word4], ignore_index=True)

# Save the combined DataFrame to a CSV file
combined_df.to_csv("crime_neighbors_race.csv", index=False)

# Load your Word2Vec model
model = Word2Vec.load("intermediate/healthcare.model")

# Extract neighbors and store in DataFrame
neighbors_df_word1 = extract_neighbors_to_dataframe(model, "racialization")
neighbors_df_word2 = extract_neighbors_to_dataframe(model, "blck")
neighbors_df_word3 = extract_neighbors_to_dataframe(model, "azn")
neighbors_df_word4 = extract_neighbors_to_dataframe(model, "latinxdad")

# Concatenate DataFrames
combined_df = pd.concat([neighbors_df_word1, neighbors_df_word2, neighbors_df_word3, neighbors_df_word4], ignore_index=True)

# Save the combined DataFrame to a CSV file
combined_df.to_csv("hc_neighbors_race.csv", index=False)
