==> code/08_mentions_all.ipynb <==
import numpy
import pickle
import csv
import os
import math
import random
import pandas as pd
import re
from gensim.models import Word2Vec

import multiprocessing as mp

csv_file_path = "intermediate/crime.csv"

df = pd.read_csv(csv_file_path, engine='python')
df = df.dropna(subset=['Year'])
df = df.loc[(df['Year']>=2004) & (df['Year']<=2022)]
df = df.drop('Unnamed: 0', axis = 1)

def mentioned(text):
    return 1 if re.search(r'\bracialization\b', text, flags=re.IGNORECASE) else 0

# Apply the function to the 'text' column to create a new column 'racialization_mentions'
df['racialization_mentions'] = df['Text'].apply(lambda x: mentioned(str(x)))

# Group by year and sum the counts of racialization mentions
grouped_counts = df.groupby('Year')['racialization_mentions'].sum()

grouped_counts.to_csv('crime_grouped_mentions.csv', header=True)

csv_file_path = "intermediate/immigration.csv"

df = pd.read_csv(csv_file_path, engine='python')
df = df.dropna(subset=['Year'])
df = df.loc[(df['Year']>=2004) & (df['Year']<=2022)]
df = df.drop('Unnamed: 0', axis = 1)

def mentioned(text):
    return 1 if re.search(r'\bracialization\b', text, flags=re.IGNORECASE) else 0

# Apply the function to the 'text' column to create a new column 'racialization_mentions'
df['racialization_mentions'] = df['Text'].apply(lambda x: mentioned(str(x)))

# Group by year and sum the counts of racialization mentions
grouped_counts = df.groupby('Year')['racialization_mentions'].sum()

grouped_counts.to_csv('immigration_grouped_mentions.csv', header=True)

csv_file_path = "intermediate/healthcare.csv"

df = pd.read_csv(csv_file_path, engine='python')
df = df.dropna(subset=['Year'])
df = df.loc[(df['Year']>=2004) & (df['Year']<=2022)]
df = df.drop('Unnamed: 0', axis = 1)

def mentioned(text):
    return 1 if re.search(r'\bracialization\b', text, flags=re.IGNORECASE) else 0

# Apply the function to the 'text' column to create a new column 'racialization_mentions'
df['racialization_mentions'] = df['Text'].apply(lambda x: mentioned(str(x)))

# Group by year and sum the counts of racialization mentions
grouped_counts = df.groupby('Year')['racialization_mentions'].sum()

grouped_counts.to_csv('healthcare_grouped_mentions.csv', header=True)
